requirements.txt
requests==2.31.0
selenium==4.15.2
python-dotenv==1.0.0
tqdm==4.66.1
faker==20.1.0
urllib3==1.26.18
PySocks==1.7.1